requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.2
cron-descriptor==1.4.5
cycler==0.12.1
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2
django-celery-beat==2.8.0
django-cloudinary-storage==0.3.0
django-crontab==0.7.1
django-extensions==4.1
django-timezone-field==7.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.58.4
fpdf==1.7.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.5.3
manage.py==0.2.10
MarkupSafe==3.0.2
matplotlib==3.10.3
msgpack==1.1.0
narwhals==1.35.0
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.0.0
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
rsa==4.9.1
runserver==1.0.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
